+++ frontend/src/features/quiz/index.ts
@@ -1,3 +1,4 @@
 export * from './QuizForm'
 export * from './QuizFormModalButton'
 export * from './useQuiz'
+export * from './types'
