+++ frontend/src/features/quiz/types.ts
@@ -1,0 +1,5 @@
+import { z } from 'zod'
+
+import { quizFormRequestSchema } from '@/libs/validates'
+
+export type QuizFormRequest = z.infer<typeof quizFormRequestSchema>
