+++ frontend/src/pages/api/quizzes/[id].ts
@@ -1,7 +1,7 @@
 import type { NextApiRequest, NextApiResponse } from 'next'
 
 import prisma from '@/libs/prisma'
-import { apiValidation, quizSchema } from '@/libs/validates'
+import { apiValidation, quizRequestSchema } from '@/libs/validates'
 
 export default async function handler(
   req: NextApiRequest,
@@ -26,7 +26,7 @@ export default async function handler(
       break
 
     case 'PUT':
-      apiValidation(req, res, quizSchema, async () => {
+      apiValidation(req, res, quizRequestSchema, async () => {
         const updatedArticle = await prisma.quiz.update({
           where: {
             id: id,
